Fix typo at the name
